Sales insights
Extracted data from SQL dumb,  cleaned and transformed the data, and then loaded it into Power BI.
Build this simple and beautiful interactive dashboard to provide sales revenue through various filters.
This will definitely give you an idea of the revenue that the company acquired over a period of time. This can be helpful to study the market trend and easily.